Add PKCE support to OAuth authorization code flow
